2.0: Changes the loop implementation
Changed the loop to use waitForKeyElements and added a check for the "more" button being visible to prevent spam clicking. 
This would originally prevent the user from opening menus on videos because of the window detecting a click, which auto-closed menus. 
The function must return true for waitForKeyElements to run it multiple times.