3.15: Better checks for drops
Now the script is capable of reading the console by hijacking console.log to check if the drops or missions are enabled.
It very specifically checks for drops OR missions, so it's possible for it to get into a situation where there isn't a drop but there is a mission it can't complete.
I don't know how to work around that, so this is fine for now.

The 10 minute timer has been completely removed and the 10 second timer is only used for YouTube because that's the one that autopauses at random.
The 10 minute timer SHOULD be unnecessary now that the script can read chat for the enabled drops directly, which might allow it to refresh immediately when drops stop showing up. This depends on the website properly updating that it can no longer get drops, which might not happen. It's hard to test since it's entirely dependent on hour long stream cycles.

The script is also capable of going through every live stream in order instead of picking whichever link loaded first.