Fixed bug in checking every live link
In the second check through every link, it deincremented the live game count by 1 by accident because the current live link number tracker is always +1 over the current check. Checking the 0th link has the tracker at 1, which should be checked against the 1 for the live game count if there's only 1. This would make it stop 1 early if the live game count has more than 1 link.